Move all-command tests to main context creation
This captures problems in the documentation and search terms by going
through the whole working set of commands
